refactor: gatewayë에 security 설정 옮기기

+++ backend/homeggu-user/src/main/java/com/homeggu/domain/preference/dto/request/PreferenceRequest.java
@@ -7,5 +7,5 @@ public class PreferenceRequest {
     private String category;
     private String mood;
     private String action;
-    private int clickedSalesBoardId;
+    private Long clickedSalesBoardId;
 }
