+++ backend/homeggu-user/src/main/java/com/homeggu/domain/saledBoardLike/dto/request/LikeRequest.java
@@ -4,5 +4,5 @@ import lombok.Getter;
 
 @Getter
 public class LikeRequest {
-    private int salesBoardId;
+    private Long salesBoardId;
 }
